Shifted bulk of Duke Main class code into new Parser class. Parser class is used to intepret commands
